fix issues on test and remove exports from selectMultiple
