+++ editor.planx.uk/src/ui/shared/SelectMultiple.tsx
@@ -4,12 +4,10 @@ import Autocomplete, {
   AutocompleteProps,
 } from "@mui/material/Autocomplete";
 import FormControl from "@mui/material/FormControl";
-import { inputLabelClasses } from "@mui/material/InputLabel";
-import { outlinedInputClasses } from "@mui/material/OutlinedInput";
 import { styled } from "@mui/material/styles";
-import TextField from "@mui/material/TextField";
 import React from "react";
-import { borderedFocusStyle } from "theme";
+
+import { StyledTextField } from "./Autocomplete/styles";
 
 const PopupIcon = (
   <ArrowIcon
@@ -54,62 +52,6 @@ const StyledAutocomplete = styled(Autocomplete)(({ theme }) => ({
     },
   },
 })) as typeof Autocomplete;
-
-export const StyledTextField = styled(TextField)(({ theme }) => ({
-  "&:focus-within": {
-    ...borderedFocusStyle,
-    [`& .${outlinedInputClasses.notchedOutline}`]: {
-      border: "1px solid transparent !important",
-    },
-  },
-  [`& .${outlinedInputClasses.notchedOutline}`]: {
-    borderRadius: 0,
-    border: `1px solid${theme.palette.border.main} !important`,
-  },
-  "& fieldset": {
-    borderColor: theme.palette.border.main,
-  },
-  backgroundColor: theme.palette.background.paper,
-  [`& .${outlinedInputClasses.root}, input`]: {
-    cursor: "pointer",
-    backgroundColor: theme.palette.background.default,
-  },
-  [`& .${inputLabelClasses.root}`]: {
-    textDecoration: "underline",
-    color: theme.palette.primary.main,
-    "&[data-shrink=true]": {
-      textDecoration: "none",
-      color: theme.palette.text.primary,
-      paddingY: 0,
-      transform: "translate(0px, -22px) scale(0.85)",
-    },
-  },
-}));
-
-export const CustomCheckbox = styled("span")(({ theme }) => ({
-  display: "inline-flex",
-  flexShrink: 0,
-  position: "relative",
-  width: 40,
-  height: 40,
-  borderColor: theme.palette.text.primary,
-  border: "2px solid",
-  background: "transparent",
-  marginRight: theme.spacing(1.5),
-  "&.selected::after": {
-    content: "''",
-    position: "absolute",
-    height: 24,
-    width: 12,
-    borderColor: theme.palette.text.primary,
-    borderBottom: "5px solid",
-    borderRight: "5px solid",
-    left: "50%",
-    top: "42%",
-    transform: "translate(-50%, -50%) rotate(45deg)",
-    cursor: "pointer",
-  },
-}));
 
 export function SelectMultiple<T>(props: Props<T>) {
   // MUI doesn't pass the Autocomplete value along to the TextField automatically
